app/pyproject.toml
[tool.poetry]
name = "payout-service"
version = "0.1.0"
description = "Service manage payout"
authors = ["Renat <ren484@mail.ru>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.2.7"
celery = "5.3.4"
redis = "5.0.1"
djangorestframework = "3.14.0"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.0"
black = "^25.12.0"
flake8 = "^7.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88